Enhance role management and order handling logic

+++ src/services/products.service.ts
@@ -1,14 +1,14 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, ForbiddenException } from '@nestjs/common';
+import { ProductsRepository } from 'src/repositories/products.repository';
 import { CreateProductDto } from 'src/dto/products/create-product.dto';
 import { UpdateProductDto } from 'src/dto/products/update-product.dto';
-import { ProductsRepository } from 'src/repositories/products.repository';
 
 @Injectable()
 export class ProductsService {
   constructor(private readonly productsRepository: ProductsRepository) {}
 
-  async create(createProductDto: CreateProductDto) {
-    return this.productsRepository.createProduct(createProductDto);
+  async create(createProductDto: CreateProductDto, supplierId: string) {
+    return this.productsRepository.createProduct(createProductDto, supplierId);
   }
 
   async findAll() {
@@ -24,7 +24,24 @@ export class ProductsService {
   }
 
   async remove(id: string) {
+    const hasActiveOrders = await this.hasActiveOrders(id);
+    if (hasActiveOrders) {
+      throw new ForbiddenException(
+        'Cannot delete product with active or delivered orders.',
+      );
+    }
     return this.productsRepository.deleteProduct(id);
+  }
+
+  async hasActiveOrders(productId: string): Promise<boolean> {
+    const orders =
+      await this.productsRepository.prismaService.orderItem.findMany({
+        where: {
+          productId,
+          order: { status: { in: ['Processing', 'Delivered'] } },
+        },
+      });
+    return orders.length > 0;
   }
 
   async disableProductsBySupplier(supplierId: string) {
